Remove in-memory FITSWriter, rename StreamedFITSWriter to FITSWriter
